Meals Dashboard v1
I have completed the first version of the Meals Dashboard. Let me know
if you have any questions about it. Also removed “newstyle.css” and put
all styles in “styles.css” — over the next day or so I’ll be going over
all of the HTML/CSS and organizing it.